+++ src/components/CameraControls.tsx
@@ -1,35 +1,9 @@
-import { OrbitControls, PerspectiveCamera } from "@react-three/drei";
-import { PerspectiveCamera as PerspectiveCameraThree, Vector3 } from "three";
-import { memo, useRef } from "react";
-import { useStore } from "@/src/types/StoreContext";
-import { observer } from "mobx-react-lite";
-import { useTravelingCamera } from "@/src/hooks/travelingCamera";
+import dynamic from "next/dynamic";
 
-type CameraControlsProps = {};
-
-export const CameraControls = memo(
-  function CameraControls({}: CameraControlsProps) {
-    // const { viewerMode } = useStore();
-
-    const cameraRef = useRef<PerspectiveCameraThree>(null);
-    const initialPositionRef = useRef(new Vector3(0, 0, 20));
-
-    useTravelingCamera(cameraRef, false);
-    // not using traveling camera because it's kinda silly but leaving around for future camera
-    // shenanigans
-    // useTravelingCamera(cameraRef, viewerMode);
-
-    return (
-      <>
-        <PerspectiveCamera
-          ref={cameraRef}
-          makeDefault
-          position={initialPositionRef.current}
-          near={0.1}
-          far={100}
-        />
-        <OrbitControls camera={cameraRef.current ?? undefined} />
-      </>
-    );
-  },
+export const CameraControls = dynamic(
+  () =>
+    import("@/src/components/CameraControlsInner").then((mod) => ({
+      default: mod.CameraControlsInner,
+    })),
+  { ssr: false },
 );
